+++ src/aui-widget/HISTORY.md
@@ -4,7 +4,7 @@
 
 ## @VERSION@
 
-No registries yet.
+* [AUI-2051](https://issues.liferay.com/browse/AUI-2051) Allows changing the value of WidgetTransition's "animated" attribute
 
 ## [3.0.2](https://github.com/liferay/alloy-ui/releases/tag/3.0.2)